# Make cpr weight calculation thread parallel