# Make CPR Weight Calculation Thread Parallel

This PR parallelizes the CPR weight calculation using OpenMP threading to improve performance, particularly important for GPU configurations where the CPU is idle and the CPU-based weight calculation can become a significant proportion of linear solve time.

## Changes

- Both `getTrueImpesWeights` and `getTrueImpesWeightsAnalytic` functions now use OpenMP parallel loops over element chunks
- Parallelized quasi-IMPES weight calculation using OpenMP over matrix rows

## Performance Impact

Testing on the first month of the 2M cell Sleipner case shows significant speedups:
- **cpr_trueimpes**: Linear solve time reduced from 9.74s to 8.33s (~15% reduction)
- **cpr_trueimpesanalytic**: Linear solve time reduced from 8.64s to 7.73s (~11% reduction)

## Context

This change was extracted from the larger CPR-AMG GPU implementation in https://github.com/OPM/opm-simulators/pull/6308 for easier review and can be merged independently to benefit both CPU and GPU linear solver configurations.