@jakobtorben 2025-08-22
This part I was unsure about, so any input here would be helpful

```cpp
// Compute storage using model's linearizer
// TODO: Remove the use of the threadded local linearizer
#ifdef _OPENMP
model.localLinearizer(omp_get_thread_num()).localResidual().computeStorage(storage, localElemCtx, /*spaceIdx=*/0, /*timeIdx=*/0);
#else
model.localLinearizer(0).localResidual().computeStorage(storage, localElemCtx, /*spaceIdx=*/0, /*timeIdx=*/0);
#endif
```