@jakobtorben 2025-08-22
jenkins build this serial rocm hipify please